+++ src/edu/jhu/hltcoe/gridsearch/rlt/filter/MaxNumRltRowFilter.java
@@ -34,7 +34,7 @@ public class MaxNumRltRowFilter implements RltRowFilter {
     }
 
     @Override
-    public void init(Rlt rlt, int numUnfilteredRows) throws IloException {
+    public void init(Rlt rlt, long numUnfilteredRows) throws IloException {
         this.initProp = initMax / numUnfilteredRows;
     }
