+++ src/edu/jhu/hltcoe/gridsearch/rlt/filter/RandPropRltRowFilter.java
@@ -20,7 +20,7 @@ public class RandPropRltRowFilter implements RltRowFilter {
     }
     
     @Override
-    public void init(Rlt rlt, int numUnfilteredRows) throws IloException {
+    public void init(Rlt rlt, long numUnfilteredRows) throws IloException {
         // Do nothing.
     }
